separate app tests into their own file
